requirements.txt
selenium
webdriver-manager
pandas
schedule
python-dotenv